+++ example/src/index.tsx
@@ -13,7 +13,6 @@ import {
   DefaultTheme,
   InitialState,
   NavigationContainer,
-  NavigatorScreenParams,
   PathConfigMap,
   useNavigationContainerRef,
 } from '@react-navigation/native';
@@ -46,117 +45,21 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
 
 import { restartApp } from './Restart';
-import AuthFlow from './Screens/AuthFlow';
-import BottomTabs from './Screens/BottomTabs';
-import DynamicTabs from './Screens/DynamicTabs';
-import LinkComponent from './Screens/LinkComponent';
-import MasterDetail from './Screens/MasterDetail';
-import MaterialBottomTabs from './Screens/MaterialBottomTabs';
-import MaterialTopTabsScreen from './Screens/MaterialTopTabs';
-import MixedHeaderMode from './Screens/MixedHeaderMode';
-import MixedStack from './Screens/MixedStack';
-import ModalStack from './Screens/ModalStack';
-import NativeStack from './Screens/NativeStack';
-import NativeStackHeaderCustomization from './Screens/NativeStackHeaderCustomization';
+import { RootDrawerParamList, RootStackParamList, SCREENS } from './screens';
 import NotFound from './Screens/NotFound';
-import PreventRemove from './Screens/PreventRemove';
-import SimpleStack from './Screens/SimpleStack';
-import StackHeaderCustomization from './Screens/StackHeaderCustomization';
-import StackTransparent from './Screens/StackTransparent';
 import SettingsItem from './Shared/SettingsItem';
 
 if (Platform.OS !== 'web') {
   LogBox.ignoreLogs(['Require cycle:']);
 }
 
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace ReactNavigation {
-    interface RootParamList extends RootStackParamList {}
-  }
-}
-
-type RootDrawerParamList = {
-  Examples: undefined;
-};
-
-const SCREENS = {
-  NativeStack: { title: 'Native Stack', component: NativeStack },
-  SimpleStack: { title: 'Simple Stack', component: SimpleStack },
-  ModalStack: {
-    title: 'Modal Stack',
-    component: ModalStack,
-  },
-  MixedStack: {
-    title: 'Regular + Modal Stack',
-    component: MixedStack,
-  },
-  MixedHeaderMode: {
-    title: 'Float + Screen Header Stack',
-    component: MixedHeaderMode,
-  },
-  StackTransparent: {
-    title: 'Transparent Stack',
-    component: StackTransparent,
-  },
-  StackHeaderCustomization: {
-    title: 'Header Customization in Stack',
-    component: StackHeaderCustomization,
-  },
-  NativeStackHeaderCustomization: {
-    title: 'Header Customization in Native Stack',
-    component: NativeStackHeaderCustomization,
-  },
-  BottomTabs: { title: 'Bottom Tabs', component: BottomTabs },
-  MaterialTopTabs: {
-    title: 'Material Top Tabs',
-    component: MaterialTopTabsScreen,
-  },
-  MaterialBottomTabs: {
-    title: 'Material Bottom Tabs',
-    component: MaterialBottomTabs,
-  },
-  DynamicTabs: {
-    title: 'Dynamic Tabs',
-    component: DynamicTabs,
-  },
-  MasterDetail: {
-    title: 'Master Detail',
-    component: MasterDetail,
-  },
-  AuthFlow: {
-    title: 'Auth Flow',
-    component: AuthFlow,
-  },
-  PreventRemove: {
-    title: 'Prevent removing screen',
-    component: PreventRemove,
-  },
-  LinkComponent: {
-    title: '<Link />',
-    component: LinkComponent,
-  },
-};
-
-type RootStackParamList = {
-  Home: NavigatorScreenParams<RootDrawerParamList>;
-  NotFound: undefined;
-} & {
-  [P in keyof typeof SCREENS]: NavigatorScreenParams<{
-    Article: { author?: string };
-    Albums: undefined;
-    Chat: undefined;
-    Contacts: undefined;
-    NewsFeed: undefined;
-    Dialog: undefined;
-  }>;
-};
-
 const Drawer = createDrawerNavigator<RootDrawerParamList>();
 const Stack = createStackNavigator<RootStackParamList>();
 
 const NAVIGATION_PERSISTENCE_KEY = 'NAVIGATION_STATE';
 const THEME_PERSISTENCE_KEY = 'THEME_TYPE';
+
+const SCREEN_NAMES = Object.keys(SCREENS) as (keyof typeof SCREENS)[];
 
 export default function App() {
   const [theme, setTheme] = React.useState(DefaultTheme);
@@ -262,9 +165,7 @@ export default function App() {
           prefixes: [createURL('/')],
           config: {
             initialRouteName: 'Home',
-            screens: (Object.keys(SCREENS) as (keyof typeof SCREENS)[]).reduce<
-              PathConfigMap<RootStackParamList>
-            >(
+            screens: SCREEN_NAMES.reduce<PathConfigMap<RootStackParamList>>(
               (acc, name) => {
                 // Convert screen names such as SimpleStack to kebab case (simple-stack)
                 const path = name
@@ -352,9 +253,11 @@ export default function App() {
                       <SafeAreaView edges={['right', 'bottom', 'left']}>
                         <SettingsItem
                           label="Right to left"
-                          value={I18nManager.isRTL}
+                          value={I18nManager.getConstants().isRTL}
                           onValueChange={() => {
-                            I18nManager.forceRTL(!I18nManager.isRTL);
+                            I18nManager.forceRTL(
+                              !I18nManager.getConstants().isRTL
+                            );
                             restartApp();
                           }}
                         />
@@ -374,20 +277,16 @@ export default function App() {
                           }}
                         />
                         <Divider />
-                        {(Object.keys(SCREENS) as (keyof typeof SCREENS)[]).map(
-                          (name) => (
-                            <List.Item
-                              key={name}
-                              testID={name}
-                              title={SCREENS[name].title}
-                              onPress={() => {
-                                // FIXME: figure this out later
-                                // @ts-expect-error
-                                navigation.navigate(name);
-                              }}
-                            />
-                          )
-                        )}
+                        {SCREEN_NAMES.map((name) => (
+                          <List.Item
+                            key={name}
+                            testID={name}
+                            title={SCREENS[name].title}
+                            onPress={() => {
+                              navigation.navigate(name);
+                            }}
+                          />
+                        ))}
                       </SafeAreaView>
                     </ScrollView>
                   )}
@@ -400,7 +299,7 @@ export default function App() {
             component={NotFound}
             options={{ title: 'Oops!' }}
           />
-          {(Object.keys(SCREENS) as (keyof typeof SCREENS)[]).map((name) => (
+          {SCREEN_NAMES.map((name) => (
             <Stack.Screen
               key={name}
               name={name}