+++ packages/drawer/src/views/modern/Drawer.tsx
@@ -258,6 +258,7 @@ export default function Drawer({
     return translateX;
   });
 
+  const isRTL = I18nManager.getConstants().isRTL;
   const drawerAnimatedStyle = useAnimatedStyle(() => {
     const distanceFromEdge = dimensions.width - drawerWidth;
 
@@ -273,10 +274,10 @@ export default function Drawer({
                   // The drawer stays in place when `drawerType` is `back`
                   (drawerType === 'back' ? 0 : translateX.value) +
                   (drawerPosition === 'left'
-                    ? I18nManager.isRTL
+                    ? isRTL
                       ? -distanceFromEdge
                       : 0
-                    : I18nManager.isRTL
+                    : isRTL
                     ? 0
                     : distanceFromEdge),
               },