+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.4](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@6.4.3...@react-navigation/drawer@6.4.4) (2022-08-24)
+
+
+### Bug Fixes
+
+* add helpful error when using legacy implementation with Reanimated v3 ([#10693](https://github.com/react-navigation/react-navigation/issues/10693)) ([d19987b](https://github.com/react-navigation/react-navigation/commit/d19987b0574975df434aa54e0f1c1fba62e7aaa3))
+* fix isRTL check in drawer ([2d2f4e3](https://github.com/react-navigation/react-navigation/commit/2d2f4e3a53a08b02866780e880338bf347092ff2))
+* replace deprecated I18nManager.isRTL with 18nManager.getConstants().isRTL ([#10547](https://github.com/react-navigation/react-navigation/issues/10547)) ([50b88d4](https://github.com/react-navigation/react-navigation/commit/50b88d40496a04f613073c63119b21a104ec9bc2))
+
+
+
+
+
 ## [6.4.3](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@6.4.2...@react-navigation/drawer@6.4.3) (2022-07-05)